move ellipse.pde into folder, minor changes
there were some division errors afoot! not so much division errors but errors in where i am/am not dividing. i was using the diameter of the ellipse not the radius :O

in any case, this produces some weird result for non-circular ellipses. i am in a lecture right now so i can't do much :( off to you, future me